Refactor: Separate play and pause methods
The play pause method will check current status of the player and then
call the opposite method of the current status. For example if the
current status is paused, the method will call the play method. The same
is true for the player if the current status is playing, it will pause.

The reason the methods were seperate is to allow the methods to be
called by other methods.  For example, when the application loads, it
could call the play method to start a song and ensure the ImageButton is
set to the correct image.